# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger Petstore",
    "description": "A sample API that uses a petstore as an example to demonstrate features in the swagger-2.0 specification",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Swagger API Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth-token-create": {
      "post": {
        "description": "Creates A new Token For Api Access",
        "operationId": "gettoken",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "createtoken",
            "in": "body",
            "description": "Create Token For Api Access",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewPet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "pet response",
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
    "/products/add": {
      "post": {
        "description": "Add Product",
        "operationId": "addproduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "addproduct",
            "in": "body",
            "description": "create a product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addnewproduct"
            }
          },
          {
            "name": "Token-Access",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Add Product Response",
            "schema": {
              "$ref": "#/definitions/addnewproduct"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
   "/products": {
      "get": {
        "summary": "Get All Products",
        "description": "returns the product list",
        "operationId": "getProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Token-Access",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "get products",
            "schema": {
              "type": "object"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/products/delete": {
      "post": {
        "description": "delete Product",
        "operationId": "delete product",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "deleteproduct",
            "in": "body",
            "description": "delete a product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/deleteproducts"
            }
          },
          {
            "name": "Token-Access",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete Product Response",
            "schema": {
              "$ref": "#/definitions/deleteproduct"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
    "/pets/{id}": {
      "get": {
        "description": "Returns a user based on a single ID, if the user does not have access to the pet",
        "operationId": "findPetById",
        "produces": [
          "application/json",
          "application/xml",
          "text/xml",
          "text/html"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to fetch",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "pet response",
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      },
      "deletebackup": {
        "description": "deletes a single pet based on the ID supplied",
        "operationId": "deletePet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "pet deleted"
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pet"
        ],
        "summary": "Deletes a pet",
        "description": "",
        "operationId": "deletePet",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "petId",
            "in": "path",
            "description": "Pet id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    }
  },
  "definitions": {
    "CreateToken": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "api_access_token": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "AddProduct": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "price": {
          "type": "string"
        }
      }
    },

    "DeleteProduct": {
      "type": "object",
      "required": [
        "productId"
      ],
      "properties": {
        "productId": {
          "type": "string"
        }
       
      }
    },
    "ErrorModel": {
      "type": "object",
      "required": [
        "code",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}